tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-iteration-count_one_value.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-iteration-count, one value</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-iteration-count-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="2 is a valid animation-iteration-count">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes fade {
         0% {
            background-color: blue;
         }
         100% {
            background-color: white;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         animation-duration: 1s;
         animation-iteration-count: 2;
      }
      .test {
         animation-name: fade;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         function aboutEqual(observed, actual) {
            var maxDeviation = 0.1;
            return (Math.abs(observed - actual) < maxDeviation);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 2500);
         div.addEventListener('animationend', function(evt) {
            clearTimeout(timeout);
            var passed = aboutEqual(evt.elapsedTime, 2);
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should gradually turn white and then abruptly turn
   blue again, two times.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
